socialNetwork/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-filter==24.3
django-import-export==4.3.3
django-simple-history==3.7.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Pillow==11.0.0
flake8==7.1.1
ruff==0.8.0
openpyxl==3.1.5
tablib[xls,xlsx]==3.7.0
